refactor(service-log): switch consumers to presentation/tanstack imports
Also renames serviceLogsByCarId references to byCarId in use-car-costs.ts
and its test, matching the new key name.
