+++ src/module/car/service-log/presentation/ui/sections/car-costs/use-car-costs.ts
@@ -1,8 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 import { useEffect } from 'react';
 
-import { queryKeys } from '@/car/service-log/infrastructure/tanstack/query/keys';
-import { getServiceLogsByCarIdQueryOptions } from '@/car/service-log/infrastructure/tanstack/query/options';
+import { queryKeys } from '@/car/service-log/presentation/tanstack/query/keys';
+import { getServiceLogsByCarIdQueryOptions } from '@/car/service-log/presentation/tanstack/query/options';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
 import { queryKeySerialize } from '@/common/presentation/tanstack/query-key';
 
@@ -22,7 +22,7 @@ export function useCarCostsSection({ carId }: UseCarCostsSectionParams) {
       addToast(
         error?.message || 'Cannot get service logs costs.',
         'error',
-        queryKeySerialize(queryKeys.serviceLogsByCarId(carId)),
+        queryKeySerialize(queryKeys.byCarId(carId)),
       );
   }, [isError, error, addToast, carId]);
